requirements.txt
# Core dependencies
dash==2.14.2
PyGithub==2.1.1
openai==1.6.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
pandas==2.1.3
plotly==5.18.0
tenacity==8.2.3
loguru==0.7.2

# Data processing and visualization
pandas==2.1.3
plotly==5.18.0

# Async support
aiohttp==3.9.1

# Utilities
tenacity==8.2.3
loguru==0.7.2

# Development tools
black==23.11.0
pylint==3.0.2
pytest==7.4.3
pytest-asyncio==0.21.1